\name{searchIkea}
\alias{searchIkea}
\title{scrape the Ikea search reseults}
\usage{
search("stoeltje")
}
\description{
This function goes to the ikea search page with given string as query, it returns a data set with all the results and dwonloads the corresponding images.
}
\examples{
stoel = ikeaSearch("stoeltjes")
}
